refactor: split airtable data into files, refactor fetch depth
