+++ packages/skylarktv/src/mocks/airtable/index.ts
@@ -1,0 +1,19 @@
+// Re-export everything from the modular airtable files
+
+// Utils and types
+export * from "./utils";
+
+// Data loading and getters
+export * from "./data";
+
+// Media object conversions
+export * from "./media-objects";
+
+// Dynamic content functionality
+export * from "./dynamic-content";
+
+// Set-related functions
+export * from "./sets";
+
+// Search functionality
+export * from "./search";
